Restore old behaviour of hiding Gold Bricks for disabled locations
The client does not actually send these Gold Bricks to datastorage when
the locations are disabled.

Gold Bricks from Bonuses were fixed earlier, as part of Ridesanity.

The better way to handle out-of-logic Gold Bricks would be for the
client to update the server's datastorage with the current Gold Brick
count, then logic that requires a number of Gold Bricks can require
the count from the client to be high enough, but then put the count from
events in square brackets to indicate the out-of-logic.